backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
sqlalchemy==2.0.35
psycopg[binary]==3.2.3
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
ezdxf==1.3.4
shapely==2.0.6
numpy==2.1.2
httpx==0.27.2
